BE/pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.0.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
]
